requirements.txt
django==6.0.3
djangorestframework==3.17.1
psycopg2-binary==2.9.11
django-filter==25.2
djangorestframework-simplejwt==5.5.1
pyjwt==2.12.1
drf-yasg==1.21.15